requirements.txt
broadlink
pySqsListener